Initial support for NodeMCU (ESP8266 WiFi library), tested on LoLin ESP3866, needs further work on #defines
