+++ libraries/SuplaDevice/log.cpp
@@ -11,7 +11,7 @@
 #include <stdarg.h>
 #include <stdio.h>
 
-#ifdef __AVR__
+#ifdef defined(ESP8266) || defined(__AVR__)
     #include <Arduino.h>
 #else
 	#include <unistd.h>
@@ -26,9 +26,15 @@
 #include "log.h"
 
 #ifdef ESP8266
-
+#include <Arduino.h>
+extern "C"{
 	#include <osapi.h>
+	#ifdef ARDUINO_ARCH_ESP8266
+	#include <ets_sys.h>
+	#endif
 	#include <mem.h>
+}
+
 	
 	#define malloc os_malloc
 	#define free os_free
@@ -101,11 +107,11 @@ char supla_log_string(char **buffer, int *size, va_list va, const char *__fmt) {
 
 #if defined(ESP8266) || defined(__AVR__)
 	void supla_vlog(int __pri, const char *message) {
-        #ifdef ESP8266
-		os_printf("%s\r\n", message);
-        #else
+        //#ifdef ESP8266
+//		os_printf("%s\r\n", message);
+//        #else
 		Serial.println(message);
-        #endif
+//        #endif
 	}
 #else
 void supla_vlog(int __pri, const char *message) {
